refactor(orchestration): extract GroupService
Five methods, 157 body lines, moved verbatim. 1704 -> 1570 lines. No receiver
rewrite was needed: every reference was already on the kernel, the deps, the
cancellation collaborator, or one of the five.

The plan's port list omitted config, which listGroupSummaries uses, and named
logger, which nothing uses.

cancelGroup is the most timing-fragile method in the refactor. It opens no
mutate of its own -- it delegates each task's mutation to
requestTaskCancellation, which opens one -- and it calls that collaborator
directly rather than through the facade's delegation. Task 14 established why:
the delegation's extra microtask hop reorders the detached startWorkerCancellation
chain's saveState against cancelGroup's own synchronous `group.cancelled` log,
and the `cancelGroup cancels its tasks` fixture goes red. The call, its await,
its argument object, and the synchronous work that follows it all moved
unchanged.

GroupService depends on TaskCancellationService; the edge stays one-way.
