+++ src/orchestration/orchestration-service.ts
@@ -14,6 +14,7 @@ import type {
 import { AsyncMutex } from "./async-mutex";
 import { sameCoordinatorSession, stableCoordinatorSession } from "./coordinator-identity";
 import { CoordinatorRegistryService } from "./service/coordinator-registry-service";
+import { GroupService } from "./service/group-service";
 import { HumanQuestionService } from "./service/human-question-service";
 import { NoticeDeliveryService } from "./service/notice-delivery-service";
 import { OrchestrationStateKernel } from "./service/orchestration-state-kernel";
@@ -308,6 +309,7 @@ export class OrchestrationService {
   private readonly coordinators: CoordinatorRegistryService;
   private readonly cancellation: TaskCancellationService;
   private readonly humanQuestions: HumanQuestionService;
+  private readonly groups: GroupService;
 
   constructor(private readonly deps: OrchestrationServiceDeps) {
     this.kernel = new OrchestrationStateKernel({ logger: deps.logger }, deps.stateMutex);
@@ -318,6 +320,7 @@ export class OrchestrationService {
     this.coordinators = new CoordinatorRegistryService(deps, this.kernel, this.workerSessions);
     this.cancellation = new TaskCancellationService(deps, this.kernel, this.workerSessions, this.questionFlow);
     this.humanQuestions = new HumanQuestionService(deps, this.kernel, this.workerSessions, this.questionFlow, this.cancellation);
+    this.groups = new GroupService(deps, this.kernel, this.cancellation);
   }
 
 
@@ -329,129 +332,25 @@ export class OrchestrationService {
     coordinatorSession: string;
     title: string;
   }): Promise<OrchestrationGroupRecord> {
-    if (input.coordinatorSession.trim().length === 0) {
-      throw new Error("coordinatorSession must be a non-empty string");
-    }
-    if (input.title.trim().length === 0) {
-      throw new Error("title must be a non-empty string");
-    }
-
-    const group = await this.kernel.mutate(async () => {
-      const state = await this.deps.loadState();
-      const now = this.deps.now().toISOString();
-      const groupId = this.deps.createId();
-      const nextGroup: OrchestrationGroupRecord = {
-        groupId,
-        coordinatorSession: input.coordinatorSession,
-        title: input.title.trim(),
-        createdAt: now,
-        updatedAt: now,
-      };
-
-      const groups = this.kernel.ensureGroups(state);
-      groups[groupId] = nextGroup;
-      await this.deps.saveState(state);
-
-      return { ...nextGroup };
-    });
-
-    this.kernel.logEvent("orchestration.group.created", "group created", this.kernel.groupContext(group));
-
-    return group;
+    return await this.groups.createGroup(input);
   }
 
   async getGroupSummary(input: {
     groupId: string;
     coordinatorSession: string;
   }): Promise<OrchestrationGroupSummary | null> {
-    const state = await this.deps.loadState();
-    const group = this.kernel.ensureGroups(state)[input.groupId];
-    if (!group || !sameCoordinatorSession(group.coordinatorSession, input.coordinatorSession)) {
-      return null;
-    }
-
-    return this.buildGroupSummary(
-      group,
-      Object.values(state.orchestration.tasks).filter((task) => task.groupId === input.groupId),
-    );
+    return await this.groups.getGroupSummary(input);
   }
 
   async listGroupSummaries(input: OrchestrationGroupListFilter): Promise<OrchestrationGroupSummary[]> {
-    const state = await this.deps.loadState();
-    const tasks = Object.values(state.orchestration.tasks);
-    const threshold = this.deps.config.orchestration.progressHeartbeatSeconds;
-    const now = this.deps.now().getTime();
-    const sortField = input.sort ?? "updatedAt";
-    const order = input.order ?? "desc";
-
-    return Object.values(this.kernel.ensureGroups(state))
-      .filter((group) => sameCoordinatorSession(group.coordinatorSession, input.coordinatorSession))
-      .map((group) => ({
-        group,
-        summary: this.buildGroupSummary(group, tasks.filter((task) => task.groupId === group.groupId)),
-      }))
-      .filter(({ summary }) => {
-        if (input.status === undefined) return true;
-        if (input.status === "pending") return summary.pendingApprovalTasks > 0;
-        if (input.status === "running") return summary.runningTasks > 0;
-        return summary.terminal === true;
-      })
-      .filter(({ group }) => {
-        if (input.stuck !== true) return true;
-        if (group.injectionPending !== true) return false;
-        const elapsed = (now - new Date(group.updatedAt).getTime()) / 1000;
-        return elapsed >= threshold;
-      })
-      .sort((left, right) => {
-        const leftValue = sortField === "createdAt" ? left.group.createdAt : left.group.updatedAt;
-        const rightValue = sortField === "createdAt" ? right.group.createdAt : right.group.updatedAt;
-        const compare = leftValue.localeCompare(rightValue);
-        return order === "asc" ? compare : -compare;
-      })
-      .map(({ summary }) => summary);
+    return await this.groups.listGroupSummaries(input);
   }
 
   async cancelGroup(input: {
     groupId: string;
     coordinatorSession: string;
   }): Promise<CancelGroupResult> {
-    const summary = await this.getGroupSummary(input);
-    if (!summary) {
-      throw new Error(`group "${input.groupId}" does not exist`);
-    }
-
-    const cancelledTaskIds: string[] = [];
-    const skippedTaskIds: string[] = [];
-
-    for (const task of summary.tasks) {
-      if (this.kernel.isTerminalStatus(task.status)) {
-        skippedTaskIds.push(task.taskId);
-        continue;
-      }
-
-      await this.cancellation.requestTaskCancellation({
-        taskId: task.taskId,
-        coordinatorSession: input.coordinatorSession,
-      });
-      cancelledTaskIds.push(task.taskId);
-    }
-
-    const refreshed = await this.getGroupSummary(input);
-    if (!refreshed) {
-      throw new Error(`group "${input.groupId}" does not exist`);
-    }
-
-    this.kernel.logEvent("orchestration.group.cancelled", "group cancelled", {
-      ...this.kernel.groupContext(refreshed.group),
-      cancelled_count: cancelledTaskIds.length,
-      skipped_count: skippedTaskIds.length,
-    });
-
-    return {
-      summary: refreshed,
-      cancelledTaskIds,
-      skippedTaskIds,
-    };
+    return await this.groups.cancelGroup(input);
   }
 
   async requestDelegate(input: RequestDelegateInput): Promise<RequestDelegateResult>;
@@ -1523,39 +1422,6 @@ export class OrchestrationService {
    */
   async reserveLogicalTransportSession(transportSession: string): Promise<() => Promise<void>> {
     return await this.workerSessions.reserveLogicalTransportSession(transportSession);
-  }
-
-  private buildGroupSummary(
-    group: OrchestrationGroupRecord,
-    tasks: OrchestrationTaskRecord[],
-  ): OrchestrationGroupSummary {
-    const sortedTasks = tasks
-      .slice()
-      .sort((left, right) => left.createdAt.localeCompare(right.createdAt))
-      .map((task) => ({ ...task }));
-
-    const pendingApprovalTasks = sortedTasks.filter(
-      (task) => task.status === "needs_confirmation",
-    ).length;
-    const runningTasks = sortedTasks.filter((task) => task.status === "running").length;
-    const completedTasks = sortedTasks.filter((task) => task.status === "completed").length;
-    const failedTasks = sortedTasks.filter((task) => task.status === "failed").length;
-    const cancelledTasks = sortedTasks.filter((task) => task.status === "cancelled").length;
-
-    return {
-      group: { ...group },
-      tasks: sortedTasks,
-      totalTasks: sortedTasks.length,
-      pendingApprovalTasks,
-      runningTasks,
-      completedTasks,
-      failedTasks,
-      cancelledTasks,
-      terminal:
-        sortedTasks.length > 0 &&
-        sortedTasks.every((task) => task.reviewPending === undefined) &&
-        sortedTasks.every((task) => this.kernel.isTerminalStatus(task.status)),
-    };
   }
 
   private resolveRpcSourceContext(
